Split daily_tasks.sh up into independent cron jobs
This gets rid of some warts, like the weekly mail_log update in the
*daily* script. More importantly, it allows users to specify independent
cadences for each of these operations, which is definitely something
which I at least want.

Note that commit does drop the sets of LANGUAGE/LC_ALL/LANG/LC_TYPE
currently present in daily_tasks.sh. I'm assuming that this is nbd, but
if it is I expect that to come up during review :)